refactor: divide common utils file into specialised files
