+++ .cproject
@@ -46,7 +46,9 @@
 								</option>
 								<option IS_BUILTIN_EMPTY="false" IS_VALUE_EMPTY="false" id="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.c.compiler.option.includepaths.43027818" name="Include paths (-I)" superClass="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.c.compiler.option.includepaths" useByScannerDiscovery="false" valueType="includePath">
 									<listOptionValue builtIn="false" value="../Core/Inc"/>
+									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/utils/Inc"/>
 									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/EEPROM/Inc"/>
+									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/utils/Inc"/>
 									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/CLI/Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32F4xx_HAL_Driver/Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32F4xx_HAL_Driver/Inc/Legacy"/>
@@ -134,7 +136,9 @@
 								</option>
 								<option IS_BUILTIN_EMPTY="false" IS_VALUE_EMPTY="false" id="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.c.compiler.option.includepaths.1338649545" name="Include paths (-I)" superClass="com.st.stm32cube.ide.mcu.gnu.managedbuild.tool.c.compiler.option.includepaths" useByScannerDiscovery="false" valueType="includePath">
 									<listOptionValue builtIn="false" value="../Core/Inc"/>
+									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/utils/Inc"/>
 									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/EEPROM/Inc"/>
+									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/utils/Inc"/>
 									<listOptionValue builtIn="false" value="../Core/CLI_EEPROM/CLI/Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32F4xx_HAL_Driver/Inc"/>
 									<listOptionValue builtIn="false" value="../Drivers/STM32F4xx_HAL_Driver/Inc/Legacy"/>
